Add easy selection of variable
Add a wrangle that selects an attribute to be the attribute used for coloring the counties. Use Attribute Promote nodes to find min and max values of the attribute. Adjust the Color node to use these min and max values instead of having to manually enter the min and max values.